Ohhhh boy a lot of thing damn

+++ src/utils/AppConfig.java
@@ -24,9 +24,9 @@ public class AppConfig {
 
     public static final int NUMBER_CHILDREN_COMMENTS = 5;
 
-    public static final int UPVOTE_SCORE_VALUE = 2;
+    public static final long UPVOTE_SCORE_VALUE = 2;
 
-    public static final int DOWNVOTE_SCORE_VALUE = -3;
+    public static final long DOWNVOTE_SCORE_VALUE = -3;
 
-    public static final int COMMENT_SCORE_VALUE = -3;
+    public static final long COMMENT_SCORE_VALUE = -3;
 }
